Write initial content
Summary:
Explain who we are, what we do, what are our values.

The explanation is a deck of cards, with an horizontal scrollbar
to navigate between them.

We use Sly (with bugfixes for non list items).

As a follow-up commit, calls for actions and links to the relevant
resources will be interesting.

Eglide logo is by @rama.

Cards follow the colors defined in M11.

Ref. T1148

Test Plan: Scroll in different browsers, check responsive breakpoint works.

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: rama

Differential Revision: https://devcentral.nasqueron.org/D904